Updated Twitter short link handling
- Short URLs on Twitter.com (behind twitter's own t.co short links) are
now recursively expanded; e.g. if a bit.ly link was shared on Twitter,
it was obfuscated behind a t.co link, and this extension would resolve
the t.co link and show you the obfuscated short bit.ly link instead of
the source of the bit.ly link. TODO: make short url expansion properly
recursive to deal with all manner of short URLs behind short URLs (to a
depth of, e.g., 3).
- Updated the default theme for new users to "Light" theme (in keeping
with modern trends for white/light coloured websites, and ensuring the
Clear Links popup is not obtrusive)